Change Renderer to a more data-oriented design, add color to test cube

+++ GraphicsToolkit/src/Camera.cpp
@@ -86,5 +86,5 @@ float Camera::GetFarClipDistance() const
 void Camera::SetFrameSize(const Vec2i& frameSize)
 {
 	this->projectionMatrixIsDirty = true;
-	this->aspectRatio = frameSize.x / frameSize.y;
+	this->aspectRatio = frameSize.x / static_cast<float>(frameSize.y);
 }
